refactor(memory-core): migrate services to SDK flatter structure (#11224)
- Relocate services from mcp/server/memory-core/services to ai/services/memory-core/
- Introduce McpIntegrationService to decouple Server.mjs
- Update SDK exports in ai/services.mjs while preserving namespace prefixes
- Shrink Server.mjs to < 50 LOC of functional code
- Address circular import in MailboxService via dynamic ESM import
